refactor(phase5): extract resource_impl_tokens from codegen.rs
Moves the 2,622-line `resource_impl_tokens` function into its own
submodule `codegen/resource_impl.rs`, reducing `codegen.rs` from
4,336 lines to 1,718 lines (−60%). All 233 passing tests preserved.
